Move ssh related functions into ssh.go.
